requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
numpy==1.24.2
packaging==23.0
pydantic==1.10.5
python-dotenv==1.0.0
python-status==1.0.1
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
